Move .raw loader plugin
