+++ source/gloperate/CMakeLists.txt
@@ -80,6 +80,7 @@ set(headers
     ${include_path}/base/ColorGradientPreparation.h
     ${include_path}/base/ExplicitEquidistantColorGradient.h
     ${include_path}/base/ExplicitEquidistantColorGradient.inl
+    ${include_path}/base/FileNameSuffix.h
 
     ${include_path}/pipeline/PipelineContainer.h
     ${include_path}/pipeline/Stage.h
@@ -153,6 +154,7 @@ set(headers
     ${include_path}/tools/ImageExporter.h
 
     ${include_path}/loaders/ShaderLoader.h
+    ${include_path}/loaders/GlrawTextureLoader.h
 )
 
 set(sources
@@ -185,6 +187,7 @@ set(sources
     ${source_path}/base/ExplicitEquidistantColorGradient.cpp
     ${source_path}/base/AbstractColorGradient.cpp
     ${source_path}/base/ExtendedProperties.cpp
+    ${source_path}/base/FileNameSuffix.cpp
 
     ${source_path}/pipeline/PipelineContainer.cpp
     ${source_path}/pipeline/Stage.cpp
@@ -245,6 +248,7 @@ set(sources
     ${source_path}/tools/ImageExporter.cpp
 
     ${source_path}/loaders/ShaderLoader.cpp
+    ${source_path}/loaders/GlrawTextureLoader.cpp
 )
 
 # Group source files
