+++ source/gloperate/include/gloperate/base/FileNameSuffix.h
