--- source/plugins/gloperate-glraw-loaders/plugin.cpp
@@ -1,11 +1,0 @@
-
-#include <cppexpose/plugin/plugin_api.h>
-
-#include "GlrawTextureLoader.h"
-
-
-CPPEXPOSE_PLUGIN_LIBRARY
-
-    CPPEXPOSE_PLUGIN_COMPONENT(GlrawTextureLoader)
-
-CPPEXPOSE_PLUGIN_LIBRARY_END
